All blog pages should now be consistent and non-horrible.
Would like to get a designer's input on how to make the bylines for both standalone blog articles web pages and the blog entries in a list look better.

For standalone pages, I just have italicized for now.

For blog articles in lists, I would want the byline to be right below the title, but I could easily get the link text to not have a big space underneath it by default.

So I'm leaving both as-is for now, at least they're halfway decent and consistent, but I'd like to make them better.

Also note that the /blog/, /blog/archive/, and /blog/2015/ pages have no direct links from anywhere else on the site right now. Would like to fix that as well, but it's not crucial.