Tweak meeting web page.
Added more text to what's shown for each chapter, so that the chapters themselves don't need as much text (and can in theory be used in different formats in the future). Specifically, now there are "Location:", "Day and Frequency:", and "Time:" labels, so the actual chapter text doesn't need to say "Location varies", it can just say "Varies" for the location field.